requirements.txt
idna==2.8
anyio==4.2.0
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2023.5.7
exceptiongroup==1.1.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
python-telegram-bot==20.7
pytz==2023.3
six==1.16.0
sniffio==1.3.0
tornado==6.3.2
tzlocal==5.0.1
pyyaml
requests==2.22.0
beautifulsoup4==4.12.2
python-dotenv==1.1.0
